feat: `msgsSlice` store slice
