+++ src/components/messages/MsgList/index.tsx
@@ -2,7 +2,7 @@ import type { ReactElement } from 'react'
 
 import { TxListGrid } from '@/components/transactions/TxList'
 import MsgListItem from '@/components/messages/MsgListItem'
-import type { MessageListPage } from '@/hooks/useMessages'
+import type { MessageListPage } from '@/store/msgsSlice'
 
 const MsgList = ({ items }: { items: MessageListPage['results'] }): ReactElement => {
   return (
