+++ src/components/messages/MsgListItem/index.tsx
@@ -3,7 +3,7 @@ import type { ReactElement } from 'react'
 import { isMessageListDateLabel, isMessageListItem } from '@/utils/message-guards'
 import TxDateLabel from '@/components/transactions/TxDateLabel'
 import ExpandableMsgItem from '@/components/messages/MsgListItem/ExpandableMsgItem'
-import type { MessageListPage } from '@/hooks/useMessages'
+import type { MessageListPage } from '@/store/msgsSlice'
 
 const MsgListItem = ({ item }: { item: MessageListPage['results'][number] }): ReactElement | null => {
   if (isMessageListDateLabel(item)) {
