+++ src/components/messages/MsgSummary/index.tsx
@@ -5,14 +5,14 @@ import type { ReactElement } from 'react'
 import DateTime from '@/components/common/DateTime'
 import useWallet from '@/hooks/wallets/useWallet'
 import useIsWrongChain from '@/hooks/useIsWrongChain'
-import { MessageStatus } from '@/hooks/useMessages'
+import { MessageStatus } from '@/store/msgsSlice'
 import MsgType from '@/components/messages/MsgType'
 import SignMsgButton from '@/components/messages/SignMsgButton'
 import useMessageStatus from '@/hooks/useMessageStatus'
 import useIsMsgPending from '@/hooks/useIsMsgPending'
 import TxConfirmations from '@/components/transactions/TxConfirmations'
 import classNames from 'classnames'
-import type { Message } from '@/hooks/useMessages'
+import type { Message } from '@/store/msgsSlice'
 
 import txSummaryCss from '@/components/transactions/TxSummary/styles.module.css'
